+++ src/reviver.ts
@@ -1,8 +1,6 @@
 import { isObject } from './is_object';
 import { List, list, LIST_OBJECT_TYPE } from './list';
 import { Idle, idle, IDLE_OBJECT_TYPE } from './idle';
-import { Tuple, tuple, TUPLE_OBJECT_TYPE } from './tuple';
-import { Container, CONTAINER_OBJECT_TYPE, wrap } from './container';
 import {
   maybe,
   Option,
@@ -29,8 +27,6 @@ export type JSONValueTypes =
 const TYPE_TO_MONAD: Record<string, Function> = {
   [LIST_OBJECT_TYPE]: list,
   [IDLE_OBJECT_TYPE]: (value: unknown) => idle(() => value),
-  [TUPLE_OBJECT_TYPE]: (values: ReadonlyArray<unknown>) => tuple(...values),
-  [CONTAINER_OBJECT_TYPE]: wrap,
   [OPTION_SOME_OBJECT_TYPE]: maybe,
   [OPTION_NONE_OBJECT_TYPE]: maybe,
   [EITHER_LEFT_OBJECT_TYPE]: left,
@@ -47,14 +43,7 @@ const TYPE_TO_MONAD: Record<string, Function> = {
 export const reviver = (
   _key: string,
   value: JSONValueTypes | SerializabledObject<unknown>,
-):
-  | JSONValueTypes
-  | Idle<any>
-  | List<any>
-  | Option<any>
-  | Container<any>
-  | Either<any, any>
-  | Tuple<ReadonlyArray<any>> =>
+): JSONValueTypes | Idle<any> | List<any> | Option<any> | Either<any, any> =>
   isObject(value) && 'type' in value && 'value' in value
     ? TYPE_TO_MONAD[value['type'] as string]?.(value['value']) ?? value
     : value;