+++ test/pipe.spec.ts
@@ -4,7 +4,7 @@ describe('pipe', () => {
   test('should compose two functions and return function.', () => {
     const composedFn = pipe(
       (n: number) => `${n} is number`,
-      (n) => n + '!'
+      (n) => n + '!',
     );
 
     expect(composedFn(9)).toBe('9 is number!');
@@ -18,13 +18,13 @@ describe('pipe', () => {
       const composedFn: any = pipe();
 
       expect(composedFn(6, 5, 4)).toEqual([6, 5, 4]);
-    }
+    },
   );
 
   test('should compose asynchronous functions.', async () => {
     const composed = pipe(
       async (s: string) => s,
-      async (s) => parseInt(s)
+      async (s: string) => parseInt(s),
     );
 
     expect(await composed('1')).toBe(1);
@@ -39,7 +39,7 @@ describe('pipe', () => {
   it('should compose functions with one mandatory parameter and with variadic parameter', () => {
     const composed = pipe(
       (n: number) => n,
-      (...args: Array<number>) => args.reduce((a, c) => a + c, 0)
+      (...args: Array<number>) => args.reduce((a, c) => a + c, 0),
     );
 
     expect(composed(1)).toBe(1);